add DTO

+++ vocab_cycle/src/main/java/de/ait/vocab_cycle/product/service/FlashCardService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface FlashCardService {
     List<FlashCard> findAll();
-    FlashCard findById(int id);
+    FlashCard findById(Long id);
     FlashCard save(FlashCard flashCard);
     FlashCard update(FlashCard flashCard);
     void delete(FlashCard flashCard);
